proj4/requirements.txt
python-dotenv
azure-cognitiveservices-speech